+++ Source/Core/joinUrls.js
@@ -78,6 +78,12 @@ define([
                 url += first.path + second.path;
             }
         } else {
+            url += first.path.replace(/^\/?/g, '');
+
+            if (appendSlash && second.path !== '') {
+                    url = url.replace(/\/?$/, '/');
+            }
+
             url += second.path;
         }